pyproject.toml
[project]
name = "lab9"
version = "0.1.0"
description = "ROS 2 Visual Explorer 后端（FastAPI + DeepSeek 流式代理）"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "httpx>=0.28",
    "python-dotenv>=1.0",
    "pydantic>=2.10",
    "flagembedding>=1.3.0,<1.5",
    "qdrant-client>=1.12",
    "tqdm>=4.66",
    "numpy>=1.26",
    "peft>=0.13",
    # 兼容性锁定：transformers 5.x 移除了 FlagReranker 依赖的 prepare_for_model；
    # FlagEmbedding 1.4 使用了 5.x 的 `dtype=` 参数，由 rag_common.py 的 monkey-patch 兼容。
    "transformers>=4.46,<4.50",
    "tokenizers>=0.20,<0.22",
    "torch>=2.5",
]

# 让 uv 从 PyTorch 官方 CUDA 索引拉 torch（cu128，匹配 NVIDIA 驱动 ≥ 12.8）。
# 否则 `uv sync` 会从 PyPI 默认源拿到 +cpu 变体，CUDA 不可用。
[tool.uv.sources]
torch = { index = "pytorch-cu128" }

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true

# ===== Ruff: lint + format（pre-commit 钩子复用此配置） =====
[tool.ruff]
target-version = "py312"
line-length = 100
extend-exclude = ["database", ".venv", "node_modules"]

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "B",    # bugbear
    "UP",   # pyupgrade
    "SIM",  # simplify
]
ignore = [
    "E501",   # 行长由 formatter 处理
    "E741",   # 数据清洗脚本里 `l` 作行变量是惯例
    "B008",   # FastAPI 依赖注入用 Depends() 作默认值
    "B904",   # 暂不强制 raise ... from（业务代码已显式抛错信息）
    "SIM103", # 多 if/return 在过滤函数中比合并的布尔表达式可读
    "SIM117", # httpx stream 等场景嵌套 with 是常见且更清晰的写法
]

[tool.ruff.lint.per-file-ignores]
# rag_common.py 必须在 import 前对 transformers 做 monkey-patch
"scripts/rag_common.py" = ["E402"]
# run_backend.py 须在 import torch 前完成 sys.path / DLL 目录设置
"run_backend.py" = ["E402"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
